requirements.txt
Django>=5.2.1
jieba>=0.42.1
pandas>=2.0.0
mysqlclient>=2.1.0
PyMySQL>=1.0.0
python-dotenv>=1.0.0
scikit-learn>=1.0.0
numpy>=1.23.0
Pillow>=9.5.0
pdfminer.six>=20221105
python-docx>=0.8.11
opencv-python>=4.7.0
torch>=2.0.0
torchvision>=0.15.0
matplotlib>=3.7.0
wordcloud>=1.9.0